+++ src/test/kotlin/tasks/MeasureAndroidLibrarySizeTaskTest.kt
@@ -4,6 +4,7 @@ import com.twilio.apkscale.tasks.MeasureAndroidLibrarySizeTask
 import junit.framework.Assert.assertEquals
 import junitparams.JUnitParamsRunner
 import junitparams.Parameters
+import org.gradle.api.artifacts.DependencySet
 import org.gradle.testfixtures.ProjectBuilder
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -12,18 +13,19 @@ import org.junit.runner.RunWith
 class MeasureAndroidLibrarySizeTaskTest {
     private val project by lazy {
         ProjectBuilder.builder()
-                .build()
+            .build()
     }
     private val abis = mutableSetOf<String>()
     private var testNdkVersion: String? = null
     private val measureAndroidLibrarySizeTask: MeasureAndroidLibrarySizeTask by lazy {
         project.tasks.create(MeasureAndroidLibrarySizeTask.MEASURE_TASK_NAME,
-                MeasureAndroidLibrarySizeTask::class.java,
-                abis,
-                21,
-                29).apply {
-            this.ndkVersion = testNdkVersion
-        }
+            MeasureAndroidLibrarySizeTask::class.java,
+            abis,
+            true,
+            21,
+            29,
+            emptyMap<String, DependencySet>(),
+            testNdkVersion ?: "")
     }
 
     @Test
@@ -61,8 +63,8 @@ class MeasureAndroidLibrarySizeTaskTest {
     @Suppress("unused")
     private fun ndkVersionParameters(): Array<Any>? {
         return arrayOf(
-                arrayOf("1.2.3", "ndkVersion = \"1.2.3\""),
-                arrayOf(null, "")
+            arrayOf("1.2.3", "ndkVersion = \"1.2.3\""),
+            arrayOf(null, "")
         )
     }
 
@@ -70,9 +72,9 @@ class MeasureAndroidLibrarySizeTaskTest {
     @Suppress("unused")
     private fun apkSplitAbiParameters(): Array<Any>? {
         return arrayOf(
-                arrayOf(emptySet<String>(), ""),
-                arrayOf(setOf("arm64-v8a"), "include \"arm64-v8a\""),
-                arrayOf(setOf("arm64-v8a", "x86_64"), "include \"arm64-v8a\", \"x86_64\"")
+            arrayOf(emptySet<String>(), ""),
+            arrayOf(setOf("arm64-v8a"), "include \"arm64-v8a\""),
+            arrayOf(setOf("arm64-v8a", "x86_64"), "include \"arm64-v8a\", \"x86_64\"")
         )
     }
 
@@ -80,9 +82,9 @@ class MeasureAndroidLibrarySizeTaskTest {
     @Suppress("unused")
     private fun apkAbiSuffixParameters(): Array<Any>? {
         return arrayOf(
-                arrayOf(emptySet<String>(), "universal", "-"),
-                arrayOf(setOf("arm64-v8a"), "universal", "-universal-"),
-                arrayOf(setOf("x86_64"), "x86_64", "-x86_64-")
+            arrayOf(emptySet<String>(), "universal", "-"),
+            arrayOf(setOf("arm64-v8a"), "universal", "-universal-"),
+            arrayOf(setOf("x86_64"), "x86_64", "-x86_64-")
         )
     }
 }